refactor: Move some code
